Djikstra is difficult
I think a priority queue makes alot of sense to build here so two
structures dont have to be hold the distance and visited nodes.
It can return the closest next node to visit without traversing through
the whole list of distanceMap intersected with the unvisited set.
Also, maybe it makes more sense to create an adjacency graph out of the
initial maze. This allows the 'priority queue' to be created from just
the starting position and discover as we go.
I think i can optimize this one a bit more by inserting into the
distance graph as we go instead of initializing the whole maze into
memory.
Also I think i can remove all walls from the distance graph from which
the unvisitedSet is generated from as they shouldn't be created anyway.