Lfu in golang
Got a loooooooooong way to go :) but its really fun and i get to play
around with pointers.
Didn't have any pointer related issues (ty lsp) maybe i'm playing in
easy mode with it on. It fixed variable assignments to address
assignments for me. Also just golang in general deals with dereferencing
for everyone which is quite nice.